water-image-convert/pyproject.toml
[tool.poetry]
name = "chlo-compute"
version = "0.0.1"
description = "Calculate chlorophyll-a concentration using SENTINEL-2 MSI bands from Earth Engine"
authors = []
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
typed-argparse = "^0.2.11"
earthengine-api = "^0.1.368"
pandas = "^2.0.3"
loguru = "^0.7.0"
numpy = "^1.25.0"
requests = "^2.31.0"
tqdm = "^4.66.1"
google-cloud-storage = "^2.10.0"
rasterio = "^1.3.8"
folium = { git = "https://github.com/python-visualization/folium.git", rev = "57e8eae" }
branca = { git = "https://github.com/python-visualization/branca.git", rev = "b6f2f21" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
img2gcs = "src.img2gcs:main"
gcs2map = "src.gcs2map:main"
